% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{audio}
\alias{audio}
\title{Add audio elements to a Shiny app}
\usage{
audio(src, type = " 'audio/mp3'", autoplay = F, controls = T)
}
\arguments{
\item{src}{source of audio file}

\item{type}{audio file extension. ie. 'audio/mp3'}

\item{autoplay}{Should the file play automatically?}

\item{controls}{Should the user control playback?}
}
\value{
None
}
\description{
Creates an audio element inside of a Shiny application
}
\examples{
audio()

}

